+++ R/bridge.R
@@ -1,3 +1,23 @@
+#' \pkg{bspm}: R Bridge to System Package Manager
+#'
+#' Enables binary package installations on Linux distributions.
+#' Provides functions to manage packages via the distribution's package
+#' manager. Also provides transparent integration with R's 'install.packages'
+#' and a fallback mechanism. When installed as a system package, interacts
+#' with the system's package manager without requiring administrative
+#' privileges via an integrated D-Bus service; otherwise, uses sudo.
+#' Currently, the following backends are supported: DNF, APT.
+#'
+#' @author Iñaki Ucar
+#'
+#' @references \url{https://github.com/Enchufa2/bspm}
+#'
+#' @docType package
+#' @name bspm-package
+#'
+#' @seealso \code{\link{manager}}, \code{\link{integration}}
+NULL
+
 utils::globalVariables(c("BUS_NAME", "OPATH", "IFACE"))
 
 backend_call <- function(method, pkgs=NULL) {