+++ R/bridge.R
@@ -1,8 +1,8 @@
-#' \pkg{bspm}: R Bridge to System Package Manager
+#' \pkg{bspm}: Bridge to System Package Manager
 #'
 #' Enables binary package installations on Linux distributions.
 #' Provides functions to manage packages via the distribution's package
-#' manager. Also provides transparent integration with R's 'install.packages'
+#' manager. Also provides transparent integration with R's \code{install.packages}
 #' and a fallback mechanism. When installed as a system package, interacts
 #' with the system's package manager without requiring administrative
 #' privileges via an integrated D-Bus service; otherwise, uses sudo.